workbench: move seed functions to separated file
